requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.93.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pynput==1.8.1
PySide6==6.8.0
PySide6_Addons==6.8.0
PySide6_Essentials==6.8.0
pywin32-ctypes==0.2.3
screeninfo==0.8.1
shiboken6==6.8.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
